TG-20: Post method finished and corrections of files and folders
