--- src/Components/Employees/AddItem/AddItem.module.css
@@ -1,7 +1,0 @@
-.container {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
