+++ src/Components/Employees/Form/EmployeesForm.js
@@ -1,0 +1,78 @@
+import React, { useState } from 'react';
+import styles from './EmployeesForm.module.css';
+
+const AddEmployee = ({ addEmployee }) => {
+  const [employeeInput, setEmployeeInput] = useState({
+    firstName: '',
+    lastName: '',
+    email: '',
+    password: ''
+  });
+  const onChange = (e) => {
+    setEmployeeInput({ ...employeeInput, [e.target.name]: e.target.value });
+  };
+  const onSubmit = (e) => {
+    e.preventDefault();
+    addEmployee(employeeInput);
+    setEmployeeInput({
+      firstName: '',
+      lastName: '',
+      email: '',
+      password: ''
+    });
+  };
+
+  return (
+    <div className={styles.container}>
+      <div>
+        <h2>Add new employee</h2>
+      </div>
+      <form onSubmit={onSubmit}>
+        <div>
+          <label>First Name</label>
+          <input
+            type="text"
+            name="firstName"
+            value={employeeInput.firstName}
+            onChange={onChange}
+            required
+          />
+        </div>
+        <div>
+          <label>Last Name</label>
+          <input
+            type="text"
+            name="lastName"
+            value={employeeInput.lastName}
+            onChange={onChange}
+            required
+          />
+        </div>
+        <div>
+          <label>Email</label>
+          <input
+            type="text"
+            name="email"
+            value={employeeInput.email}
+            onChange={onChange}
+            required
+          />
+        </div>
+        <div>
+          <label>Password</label>
+          <input
+            type="password"
+            name="password"
+            value={employeeInput.password}
+            onChange={onChange}
+            required
+          />
+        </div>
+        <input type="submit" value="Add Employee" />
+        <input type="submit" value="Edit Employee" />
+      </form>
+    </div>
+  );
+};
+
+export default AddEmployee;
